pyproject.toml
[project]
name = "javis"
version = "0.1.0"
description = "JAVIS - Personal AI Assistant with Custom Fine-tuned Model"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]

dependencies = [
    # HTTP & API
    "httpx>=0.25.0",
    "websockets>=12.0",

    # Web Server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # Data Processing
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "chromadb>=0.4.0",

    # AI/ML (Local embedding)
    "sentence-transformers>=2.2.0",

    # Utilities
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
training = [
    "unsloth>=2024.1",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.42.0",
    "peft>=0.7.0",
    "trl>=0.7.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
javis = "javis.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
